🏗️ Refactor BookReader component into modular architecture
Implement comprehensive refactoring of the monolithic BookReader component
following the four-phase plan. This transformation splits 800+ lines of
complex code into focused, maintainable components and custom hooks.

## Phase 1: Custom Hooks
- **useBookReader**: Main business logic coordinating all book-related state
- **useProgressManager**: Progress tracking and navigation controls logic
- **useTextSelection**: Text selection and dictionary lookup functionality

## Phase 2: UI Components
- **WordTooltip**: Dictionary tooltip with translation display
- **ReaderNavigation**: Bottom navigation with progress bar and page controls
- **EpubViewer**: EPUB rendering container component

## Phase 3: Modal Components
- **TableOfContents**: TOC panel for chapter navigation
- **VocabularyPanel**: Saved words management panel

## Phase 4: Main Component Refactor
- Reduced BookReader from 800+ lines to ~190 lines
- Component now acts as clean orchestrator using extracted hooks/components
- All business logic moved to custom hooks
- All UI rendering moved to focused components

## Benefits
- **Separation of Concerns**: Each component has single responsibility
- **Improved Testability**: Smaller, focused units can be tested independently
- **Better Maintainability**: Changes to one concern don't affect others
- **Enhanced Reusability**: Components and hooks can be shared/reused
- **Performance Optimization**: Individual components can be memoized

Maintains all existing functionality while dramatically improving code
organization and maintainability.
