+++ epub-reader/src/components/reader/VocabularyPanel.css
@@ -1,0 +1,81 @@
+.vocabulary-panel {
+  position: fixed;
+  top: 60px;
+  left: 20px;
+  width: 300px;
+  max-height: calc(100vh - 200px);
+  background: var(--bg-primary);
+  border: 1px solid var(--border-color);
+  border-radius: 8px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+  z-index: 200;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+}
+
+.vocabulary-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 16px;
+  border-bottom: 1px solid var(--border-color);
+  background: var(--bg-secondary);
+}
+
+.vocabulary-header h3 {
+  margin: 0;
+  font-size: 16px;
+  color: var(--text-primary);
+}
+
+.vocabulary-list {
+  flex: 1;
+  overflow-y: auto;
+  padding: 8px;
+}
+
+.no-words {
+  padding: 24px 16px;
+  text-align: center;
+  color: var(--text-secondary);
+  font-size: 14px;
+  line-height: 1.4;
+  margin: 0;
+}
+
+.vocabulary-item {
+  padding: 12px 16px;
+  border-radius: 6px;
+  margin-bottom: 8px;
+  background: var(--bg-secondary);
+  transition: background-color 0.2s;
+}
+
+.vocabulary-item:hover {
+  background: var(--border-color);
+}
+
+.word-main {
+  font-weight: bold;
+  color: var(--text-primary);
+  font-size: 16px;
+  margin-bottom: 4px;
+}
+
+.word-translation {
+  color: var(--text-secondary);
+  font-size: 14px;
+  line-height: 1.4;
+}
+
+@media (max-width: 768px) {
+  .vocabulary-panel {
+    position: fixed;
+    top: 60px;
+    left: 20px;
+    right: 20px;
+    width: auto;
+    max-height: calc(100vh - 140px);
+  }
+}
