+++ epub-reader/src/components/reader/index.ts
@@ -1,0 +1,5 @@
+export { WordTooltip, type WordTooltipProps } from './WordTooltip';
+export { ReaderNavigation, type ReaderNavigationProps } from './ReaderNavigation';
+export { EpubViewer, type EpubViewerProps } from './EpubViewer';
+export { TableOfContents, type TableOfContentsProps, type TOCItem } from './TableOfContents';
+export { VocabularyPanel, type VocabularyPanelProps } from './VocabularyPanel';
