+++ epub-reader/src/hooks/index.ts
@@ -1,0 +1,3 @@
+export { useProgressManager, type ReaderControls } from './useProgressManager';
+export { useTextSelection, type WordTooltip } from './useTextSelection';
+export { useBookReader } from './useBookReader';
